+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- (VRCSDK3) Add *Remove Vertex Color from Meshes* option to the converter window.
+- Add VertexColorRemover component in order to control removing/restoring vertex color.
+
+### Changed
+- (VRCSDK3) Vertex colors are no longer removed automatically. Use VertexColorRemover component instead.
+
 ## [1.9.1] - 2023-02-13
 
 ### Changed